Added find by ID

+++ src/main/kotlin/com/example/kotlinmsusers/repositories/UsersRepository.kt
@@ -4,4 +4,4 @@ import com.example.kotlinmsusers.models.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
-interface UsersRepository: CrudRepository<User, Int>
+interface UsersRepository: CrudRepository<User, Long>
